Brought in UserWrappr with TouchPassr-0.2.0
