+++ Source/References/UserWrappr-0.2.0.js
@@ -3331,21 +3331,21 @@ var InputWritr;
          *
          * @param {Function, String} event   The event function (or string alias of
          *                                   it) that will be called.
-         * @param {Number} [keycode]   The alias of the event function under
-         *                             triggers[event], if event is a String.
+         * @param {Mixed} [keyCode]   The alias of the event function under
+         *                            triggers[event], if event is a String.
          * @param {Event} [sourceEvent]   The raw event that caused the calling Pipe
          *                                to be triggered, such as a MouseEvent.
          * @return {Mixed}
          */
-        InputWritr.prototype.callEvent = function (event, keycode, sourceEvent) {
-            if (!this.canTrigger(event, keycode)) {
+        InputWritr.prototype.callEvent = function (event, keyCode, sourceEvent) {
+            if (!this.canTrigger(event, keyCode)) {
                 return;
             }
             if (!event) {
                 throw new Error("Blank event given, ignoring it.");
             }
             if (event.constructor === String) {
-                event = this.triggers[event][keycode];
+                event = this.triggers[event][keyCode];
             }
             return event(this.eventInformation, sourceEvent);
         };
@@ -3385,7 +3385,7 @@ var InputWritr;
         /**
          * Curry utility to create a closure that runs call() when called.
          *
-         * @param {Array} info   An array containing [alias, keycode].
+         * @param {Array} info   An array containing [alias, keyCode].
          * @return {Function} A closure Function that activates a trigger
          *                    when called.
          */
@@ -10384,6 +10384,696 @@ var ThingHittr;
     })();
     _ThingHittr.ThingHittr = ThingHittr;
 })(ThingHittr || (ThingHittr = {}));
+/// <reference path="InputWritr-0.2.0.ts" />
+var __extends = this.__extends || function (d, b) {
+    for (var p in b) if (b.hasOwnProperty(p)) d[p] = b[p];
+    function __() { this.constructor = d; }
+    __.prototype = b.prototype;
+    d.prototype = new __();
+};
+var TouchPassr;
+(function (_TouchPassr) {
+    "use strict";
+    /**
+     * Abstract class for on-screen controls. Element creation for .element
+     * and .elementInner within the constrained position is provided.
+     */
+    var Control = (function () {
+        /**
+         * Resets the control by setting member variables and calling resetElement.
+         *
+         * @param {InputWritr} InputWriter
+         * @param {Object} schema
+         */
+        function Control(InputWriter, schema, styles) {
+            this.InputWriter = InputWriter;
+            this.schema = schema;
+            this.resetElement(styles);
+        }
+        /**
+         * @return {HTMLElement} The outer container element.
+         */
+        Control.prototype.getElement = function () {
+            return this.element;
+        };
+        /**
+         * @return {HTMLElement} The inner container element.
+         */
+        Control.prototype.getElementInner = function () {
+            return this.elementInner;
+        };
+        /**
+         * Creates and returns an HTMLElement of the specified type. Any additional
+         * settings Objects may be given to be proliferated onto the Element via
+         * proliferateElement.
+         *
+         * @param {String} type   The tag of the Element to be created.
+         * @param {Object} [settings]   Additional settings for the Element, such as
+         *                              className or style.
+         * @return {HTMLElement}
+         */
+        Control.prototype.createElement = function (tag) {
+            var args = [];
+            for (var _i = 1; _i < arguments.length; _i++) {
+                args[_i - 1] = arguments[_i];
+            }
+            var element = document.createElement(tag || "div"), i;
+            for (i = 1; i < arguments.length; i += 1) {
+                this.proliferateElement(element, arguments[i]);
+            }
+            return element;
+        };
+        /**
+         * Identical to proliferate, but tailored for HTML elements because many
+         * element attributes don't play nicely with JavaScript Array standards.
+         * Looking at you, HTMLCollection!
+         *
+         * @param {HTMLElement} recipient
+         * @param {Any} donor
+         * @param {Boolean} [noOverride]
+         * @return {HTMLElement}
+         */
+        Control.prototype.proliferateElement = function (recipient, donor, noOverride) {
+            if (noOverride === void 0) { noOverride = false; }
+            var setting, i, j;
+            for (i in donor) {
+                if (donor.hasOwnProperty(i)) {
+                    // If noOverride, don't override already existing properties
+                    if (noOverride && recipient.hasOwnProperty(i)) {
+                        continue;
+                    }
+                    setting = donor[i];
+                    switch (i) {
+                        case "children":
+                            if (typeof (setting) !== "undefined") {
+                                for (j = 0; j < setting.length; j += 1) {
+                                    recipient.appendChild(setting[j]);
+                                }
+                            }
+                            break;
+                        case "style":
+                            this.proliferateElement(recipient[i], setting);
+                            break;
+                        default:
+                            // If it's null, don't do anything (like .textContent)
+                            if (setting === null) {
+                                recipient[i] = null;
+                            }
+                            else if (typeof setting === "object") {
+                                // If it's an object, recurse on a new version of it
+                                if (!recipient.hasOwnProperty(i)) {
+                                    recipient[i] = new setting.constructor();
+                                }
+                                this.proliferateElement(recipient[i], setting, noOverride);
+                            }
+                            else {
+                                // Regular primitives are easy to copy otherwise
+                                recipient[i] = setting;
+                            }
+                            break;
+                    }
+                }
+            }
+            return recipient;
+        };
+        /**
+         * Resets the container elements. In any inherited resetElement, this should
+         * still be called, as it implements the schema's position.
+         *
+         * @param {Object} styles   Container styles for the contained elements.
+         */
+        Control.prototype.resetElement = function (styles, customType) {
+            var position = this.schema.position, offset = position.offset;
+            this.element = this.createElement("div", {
+                "className": "control",
+                "style": {
+                    "position": "absolute",
+                    "width": 0,
+                    "height": 0,
+                    "boxSizing": "border-box",
+                    "opacity": ".84"
+                }
+            });
+            this.elementInner = this.createElement("div", {
+                "className": "control-inner",
+                "textContent": this.schema.label,
+                "style": {
+                    "position": "absolute",
+                    "boxSizing": "border-box",
+                    "textAlign": "center"
+                }
+            });
+            this.element.appendChild(this.elementInner);
+            if (position.horizontal === "left") {
+                this.element.style.left = "0";
+            }
+            else if (position.horizontal === "right") {
+                this.element.style.right = "0";
+            }
+            else if (position.horizontal === "center") {
+                this.element.style.left = "50%";
+            }
+            if (position.vertical === "top") {
+                this.element.style.top = "0";
+            }
+            else if (position.vertical === "bottom") {
+                this.element.style.bottom = "0";
+            }
+            else if (position.vertical === "center") {
+                this.element.style.top = "50%";
+            }
+            this.passElementStyles(styles.global);
+            this.passElementStyles(styles[customType]);
+            this.passElementStyles(this.schema.styles);
+            if (offset.left) {
+                this.elementInner.style.marginLeft = this.createPixelMeasurement(offset.left);
+            }
+            if (offset.top) {
+                this.elementInner.style.marginTop = this.createPixelMeasurement(offset.top);
+            }
+            // elementInner's center-based positioning must wait until its total width is done setting
+            setTimeout(function () {
+                if (position.horizontal === "center") {
+                    this.elementInner.style.left = Math.round(this.elementInner.offsetWidth / -2) + "px";
+                }
+                if (position.vertical === "center") {
+                    this.elementInner.style.top = Math.round(this.elementInner.offsetHeight / -2) + "px";
+                }
+            }.bind(this));
+        };
+        /**
+         * Converts a String or Number into a CSS-ready String measurement.
+         *
+         * @param {Mixed} raw   A raw measurement, such as "7" or "7px" or "7em".
+         * @return {String} The raw measurement as a CSS measurement.
+         */
+        Control.prototype.createPixelMeasurement = function (raw) {
+            if (!raw) {
+                return "0";
+            }
+            if (typeof raw === "number" || raw.constructor === Number) {
+                return raw + "px";
+            }
+            return raw;
+        };
+        /**
+         * Passes a style schema to .element and .elementInner.
+         *
+         * @param {Object} styles   A container for styles to apply.
+         */
+        Control.prototype.passElementStyles = function (styles) {
+            if (!styles) {
+                return;
+            }
+            if (styles.element) {
+                this.proliferateElement(this.element, styles.element);
+            }
+            if (styles.elementInner) {
+                this.proliferateElement(this.elementInner, styles.elementInner);
+            }
+        };
+        /**
+         * Sets the rotation of an HTML element via CSS.
+         *
+         * @param {HTMLElement} element
+         * @param {Number} rotation
+         */
+        Control.prototype.setRotation = function (element, rotation) {
+            element.style.transform = "rotate(" + rotation + "deg)";
+        };
+        /**
+         * Finds the position offset of an element relative to the page, factoring in
+         * its parent elements' offsets recursively.
+         *
+         * @param {HTMLElement} element
+         * @return {Number[]} The left and top offset of the element, in px.
+         */
+        Control.prototype.getOffsets = function (element) {
+            var output;
+            if (element.offsetParent && element !== element.offsetParent) {
+                output = this.getOffsets(element.offsetParent);
+                output[0] += element.offsetLeft;
+                output[1] += element.offsetTop;
+            }
+            else {
+                output = [element.offsetLeft, element.offsetTop];
+            }
+            return output;
+        };
+        return Control;
+    })();
+    _TouchPassr.Control = Control;
+    /**
+     * Simple button control. It activates its triggers when the users presses
+     * it or releases it, and contains a simple label.
+     */
+    var ButtonControl = (function (_super) {
+        __extends(ButtonControl, _super);
+        function ButtonControl() {
+            _super.apply(this, arguments);
+        }
+        /**
+         * Resets the elements by adding listeners for mouse and touch
+         * activation and deactivation events.
+         *
+         * @param {Object} styles   Container styles for the contained elements.
+         */
+        ButtonControl.prototype.resetElement = function (styles) {
+            var onActivated = this.onEvent.bind(this, "activated"), onDeactivated = this.onEvent.bind(this, "deactivated");
+            _super.prototype.resetElement.call(this, styles, "Button");
+            this.element.addEventListener("mousedown", onActivated);
+            this.element.addEventListener("touchstart", onActivated);
+            this.element.addEventListener("mouseup", onDeactivated);
+            this.element.addEventListener("touchend", onDeactivated);
+        };
+        /**
+         * Reation callback for a triggered event.
+         *
+         * @param {String} which   The pipe being activated, such as
+         *                         "activated" or "deactivated".
+         * @param {Event} event
+         */
+        ButtonControl.prototype.onEvent = function (which, event) {
+            var events = this.schema.pipes[which], i, j;
+            if (!events) {
+                return;
+            }
+            for (i in events) {
+                if (!events.hasOwnProperty(i)) {
+                    continue;
+                }
+                for (j = 0; j < events[i].length; j += 1) {
+                    this.InputWriter.callEvent(i, events[i][j], event);
+                }
+            }
+        };
+        return ButtonControl;
+    })(Control);
+    _TouchPassr.ButtonControl = ButtonControl;
+    /**
+     * Joystick control. An inner circle can be dragged to one of a number
+     * of directions to trigger pipes on and off.
+     */
+    var JoystickControl = (function (_super) {
+        __extends(JoystickControl, _super);
+        function JoystickControl() {
+            _super.apply(this, arguments);
+        }
+        /**
+         * Resets the element by creating a tick for each direction, along with
+         * the multiple circular elements with their triggers.
+         *
+         * @param {Object} styles   Container styles for the contained elements.
+         */
+        JoystickControl.prototype.resetElement = function (styles) {
+            _super.prototype.resetElement.call(this, styles, "Joystick");
+            var directions = this.schema.directions, element, degrees, sin, cos, dx, dy, i;
+            this.proliferateElement(this.elementInner, {
+                "style": {
+                    "border-radius": "100%"
+                }
+            });
+            // The visible circle is what is actually visible to the user
+            this.elementCircle = this.createElement("div", {
+                "className": "control-inner control-joystick-circle",
+                "style": {
+                    "position": "absolute",
+                    "background": "red",
+                    "borderRadius": "100%"
+                }
+            });
+            this.proliferateElement(this.elementCircle, styles.Joystick.circle);
+            for (i = 0; i < directions.length; i += 1) {
+                degrees = directions[i].degrees;
+                // sin and cos are an amount / 1 the tick is offset from the center
+                sin = Math.sin(degrees * Math.PI / 180);
+                cos = Math.cos(degrees * Math.PI / 180);
+                // dx and dy are measured as percent from the center, based on sin & cos
+                dx = cos * 50 + 50;
+                dy = sin * 50 + 50;
+                element = this.createElement("div", {
+                    "className": "control-joystick-tick",
+                    "style": {
+                        "position": "absolute",
+                        "left": dx + "%",
+                        "top": dy + "%",
+                        "marginLeft": (-cos * 5 - 5) + "px",
+                        "marginTop": (-sin * 2 - 1) + "px"
+                    }
+                });
+                this.proliferateElement(element, styles.Joystick.tick);
+                this.setRotation(element, degrees);
+                this.elementCircle.appendChild(element);
+            }
+            // In addition to the ticks, a drag element shows current direction
+            this.elementDragLine = this.createElement("div", {
+                "className": "control-joystick-drag-line",
+                "style": {
+                    "position": "absolute",
+                    "opacity": "0",
+                    "top": ".77cm",
+                    "left": ".77cm"
+                }
+            });
+            this.proliferateElement(this.elementDragLine, styles.Joystick.dragLine);
+            this.elementCircle.appendChild(this.elementDragLine);
+            // A shadow-like circle supports the drag effect
+            this.elementDragShadow = this.createElement("div", {
+                "className": "control-joystick-drag-shadow",
+                "style": {
+                    "position": "absolute",
+                    "opacity": "1",
+                    "top": "14%",
+                    "right": "14%",
+                    "bottom": "14%",
+                    "left": "14%",
+                    "marginLeft": "0",
+                    "marginTop": "0",
+                    "borderRadius": "100%"
+                }
+            });
+            this.proliferateElement(this.elementDragShadow, styles.Joystick.dragShadow);
+            this.elementCircle.appendChild(this.elementDragShadow);
+            this.elementInner.appendChild(this.elementCircle);
+            this.elementInner.addEventListener("click", this.triggerDragger.bind(this));
+            this.elementInner.addEventListener("touchmove", this.triggerDragger.bind(this));
+            this.elementInner.addEventListener("mousemove", this.triggerDragger.bind(this));
+            this.elementInner.addEventListener("mouseover", this.positionDraggerEnable.bind(this));
+            this.elementInner.addEventListener("touchstart", this.positionDraggerEnable.bind(this));
+            this.elementInner.addEventListener("mouseout", this.positionDraggerDisable.bind(this));
+            this.elementInner.addEventListener("touchend", this.positionDraggerDisable.bind(this));
+        };
+        /**
+         * Enables dragging, showing the elementDragLine.
+         */
+        JoystickControl.prototype.positionDraggerEnable = function () {
+            this.dragEnabled = true;
+            this.elementDragLine.style.opacity = "1";
+        };
+        /**
+         * Disables dragging, hiding the drag line and re-centering the
+         * inner circle shadow.
+         */
+        JoystickControl.prototype.positionDraggerDisable = function () {
+            this.dragEnabled = false;
+            this.elementDragLine.style.opacity = "0";
+            this.elementDragShadow.style.top = "14%";
+            this.elementDragShadow.style.right = "14%";
+            this.elementDragShadow.style.bottom = "14%";
+            this.elementDragShadow.style.left = "14%";
+            if (this.currentDirection) {
+                if (this.currentDirection.pipes && this.currentDirection.pipes.deactivated) {
+                    this.onEvent(this.currentDirection.pipes.deactivated, event);
+                }
+                this.currentDirection = undefined;
+            }
+        };
+        /**
+         * Triggers a movement point for the joystick, and snaps the stick to
+         * the nearest direction (based on the angle from the center to the point).
+         *
+         * @param {Event} event
+         */
+        JoystickControl.prototype.triggerDragger = function (event) {
+            event.preventDefault();
+            if (!this.dragEnabled) {
+                return;
+            }
+            var coordinates = this.getEventCoordinates(event), x = coordinates[0], y = coordinates[1], offsets = this.getOffsets(this.elementInner), midX = offsets[0] + this.elementInner.offsetWidth / 2, midY = offsets[1] + this.elementInner.offsetHeight / 2, dxRaw = (x - midX) | 0, dyRaw = (midY - y) | 0, thetaRaw = this.getThetaRaw(dxRaw, dyRaw), directionNumber = this.findClosestDirection(thetaRaw), direction = this.schema.directions[directionNumber], theta = direction.degrees, components = this.getThetaComponents(theta), dx = components[0], dy = -components[1];
+            this.proliferateElement(this.elementDragLine, {
+                "style": {
+                    "marginLeft": ((dx * 77) | 0) + "%",
+                    "marginTop": ((dy * 77) | 0) + "%"
+                }
+            });
+            this.proliferateElement(this.elementDragShadow, {
+                "style": {
+                    "top": ((14 + dy * 10) | 0) + "%",
+                    "right": ((14 - dx * 10) | 0) + "%",
+                    "bottom": ((14 - dy * 10) | 0) + "%",
+                    "left": ((14 + dx * 10) | 0) + "%"
+                }
+            });
+            // Ensure theta is above 0, and offset it by 90 for visual rotation
+            theta = (theta + 450) % 360;
+            this.setRotation(this.elementDragLine, theta);
+            this.positionDraggerEnable();
+            this.setCurrentDirection(direction, event);
+        };
+        /**
+         * Finds the raw coordinates of an event, whether it's a drag (touch)
+         * or mouse event.
+         *
+         * @return {Number[]} The x- and y- coordinates of the event.
+         */
+        JoystickControl.prototype.getEventCoordinates = function (event) {
+            if (event.type === "touchmove") {
+                // TypeScript 1.5 doesn't seem to have TouchEvent yet.
+                var touch = event.touches[0];
+                return [touch.pageX, touch.pageY];
+            }
+            return [event.x, event.y];
+        };
+        /**
+         * Finds the angle from a joystick center to an x and y. This assumes
+         * straight up is 0, to the right is 90, down is 180, and left is 270.
+         *
+         * @return {Number} The degrees to the given point.
+         */
+        JoystickControl.prototype.getThetaRaw = function (dxRaw, dyRaw) {
+            // Based on the quadrant, theta changes...
+            if (dxRaw > 0) {
+                if (dyRaw > 0) {
+                    // Quadrant I
+                    return Math.atan(dxRaw / dyRaw) * 180 / Math.PI;
+                }
+                else {
+                    // Quadrant II
+                    return -Math.atan(dyRaw / dxRaw) * 180 / Math.PI + 90;
+                }
+            }
+            else {
+                if (dyRaw < 0) {
+                    // Quadrant III
+                    return Math.atan(dxRaw / dyRaw) * 180 / Math.PI + 180;
+                }
+                else {
+                    // Quadrant IV
+                    return -Math.atan(dyRaw / dxRaw) * 180 / Math.PI + 270;
+                }
+            }
+        };
+        /**
+         * Converts an angle to its relative dx and dy coordinates.
+         *
+         * @param {Number} thetaRaw
+         * @return {Number[]} The x- and y- parts of an angle.
+         */
+        JoystickControl.prototype.getThetaComponents = function (thetaRaw) {
+            var theta = thetaRaw * Math.PI / 180;
+            return [Math.sin(theta), Math.cos(theta)];
+        };
+        /**
+         * Finds the index of the closest direction to an angle.
+         *
+         * @param {Number} degrees
+         * @return {Number}
+         */
+        JoystickControl.prototype.findClosestDirection = function (degrees) {
+            var directions = this.schema.directions, difference = Math.abs(directions[0].degrees - degrees), smallestDegrees = directions[0].degrees, smallestDegreesRecord = 0, record = 0, differenceTest, i;
+            for (i = 1; i < directions.length; i += 1) {
+                differenceTest = Math.abs(directions[i].degrees - degrees);
+                if (differenceTest < difference) {
+                    difference = differenceTest;
+                    record = i;
+                }
+                if (directions[i].degrees < smallestDegrees) {
+                    smallestDegrees = directions[i].degrees;
+                    smallestDegreesRecord = i;
+                }
+            }
+            // 359 is closer to 360 than 0, so pretend the smallest is above 360
+            differenceTest = Math.abs(smallestDegrees + 360 - degrees);
+            if (differenceTest < difference) {
+                difference = differenceTest;
+                record = smallestDegreesRecord;
+            }
+            return record;
+        };
+        /**
+         * Sets the current direction of the joystick, calling the relevant
+         * InputWriter pipes if necessary.
+         *
+         * @param {Object} direction
+         * @param {Event} [event]
+         */
+        JoystickControl.prototype.setCurrentDirection = function (direction, event) {
+            if (this.currentDirection === direction) {
+                return;
+            }
+            if (this.currentDirection && this.currentDirection.pipes) {
+                if (this.currentDirection.pipes.deactivated) {
+                    this.onEvent(this.currentDirection.pipes.deactivated, event);
+                }
+            }
+            if (direction.pipes && direction.pipes.activated) {
+                this.onEvent(direction.pipes.activated, event);
+            }
+            this.currentDirection = direction;
+        };
+        /**
+         * Trigger for calling pipes when a new direction is set. All children
+         * of the pipe has each of its keys triggered.
+         *
+         * @param {Object} pipes
+         * @param {Event} [event]
+         */
+        JoystickControl.prototype.onEvent = function (pipes, event) {
+            var i, j;
+            for (i in pipes) {
+                if (!pipes.hasOwnProperty(i)) {
+                    continue;
+                }
+                for (j = 0; j < pipes[i].length; j += 1) {
+                    this.InputWriter.callEvent(i, pipes[i][j], event);
+                }
+            }
+        };
+        return JoystickControl;
+    })(Control);
+    _TouchPassr.JoystickControl = JoystickControl;
+    /**
+     *
+     */
+    var TouchPassr = (function () {
+        /**
+         * @param {ITouchPassrSettings} settings
+         */
+        function TouchPassr(settings) {
+            this.InputWriter = settings.InputWriter;
+            this.styles = settings.styles || {};
+            this.resetContainer(settings.container);
+            this.controls = {};
+            if (settings.controls) {
+                this.addControls(settings.controls);
+            }
+            if (typeof settings.enabled === "undefined") {
+                this.enabled = true;
+            }
+            else {
+                this.enabled = settings.enabled;
+            }
+            this.enabled ? this.enable() : this.disable();
+        }
+        /* Simple gets
+        */
+        /**
+         * @return {InputWritr} The InputWritr for controls to pipe event triggers to.
+         */
+        TouchPassr.prototype.getInputWriter = function () {
+            return this.InputWriter;
+        };
+        /**
+         * @return {Boolean} Whether this is currently enabled and visually on the screen.
+         */
+        TouchPassr.prototype.getEnabled = function () {
+            return this.enabled;
+        };
+        /**
+         * @return {Object} The root container for styles to be added to control elements.
+         */
+        TouchPassr.prototype.getStyles = function () {
+            return this.styles;
+        };
+        /**
+         * @return {Object} The container for generated controls, keyed by their name.
+         */
+        TouchPassr.prototype.getControls = function () {
+            return this.controls;
+        };
+        /**
+         * @return {HTMLElement} The HTMLElement all controls are placed within.
+         */
+        TouchPassr.prototype.getContainer = function () {
+            return this.container;
+        };
+        /* Core functionality
+        */
+        /**
+         * Enables the TouchPassr by showing the container.
+         */
+        TouchPassr.prototype.enable = function () {
+            this.enabled = true;
+            this.container.style.display = "block";
+        };
+        /**
+         * Disables the TouchPassr by hiding the container.
+         */
+        TouchPassr.prototype.disable = function () {
+            this.enabled = false;
+            this.container.style.display = "none";
+        };
+        /**
+         * Adds any number of controls to the internal listing and HTML container.
+         *
+         * @param {Object} schemas   Schemas for new controls to be made, keyed by name.
+         */
+        TouchPassr.prototype.addControls = function (schemas) {
+            var i;
+            for (i in schemas) {
+                if (schemas.hasOwnProperty(i)) {
+                    this.addControl(schemas[i]);
+                }
+            }
+        };
+        /**
+         * Adds a control to the internal listing and HTML container.
+         *
+         * @param {Object} schema   The schema for the new control to be made.
+         */
+        TouchPassr.prototype.addControl = function (schema) {
+            var control;
+            switch (schema.control) {
+                case "Button":
+                    control = new ButtonControl(this.InputWriter, schema, this.styles);
+                    break;
+                case "Joystick":
+                    control = new JoystickControl(this.InputWriter, schema, this.styles);
+                    break;
+                default:
+                    break;
+            }
+            this.controls[schema.name] = control;
+            this.container.appendChild(control.getElement());
+        };
+        /* HTML manipulations
+        */
+        /**
+         * Resets the base controls container. If a parent element is provided,
+         * the container is added to it.
+         *
+         * @param {HTMLElement} [parentContainer]   A container element, such as
+         *                                          from GameStartr.
+         */
+        TouchPassr.prototype.resetContainer = function (parentContainer) {
+            this.container = Control.prototype.createElement("div", {
+                "className": "touch-passer-container",
+                "style": {
+                    "position": "absolute",
+                    "top": 0,
+                    "right": 0,
+                    "bottom": 0,
+                    "left": 0
+                }
+            });
+            if (parentContainer) {
+                parentContainer.appendChild(this.container);
+            }
+        };
+        return TouchPassr;
+    })();
+    _TouchPassr.TouchPassr = TouchPassr;
+})(TouchPassr || (TouchPassr = {}));
 var WorldSeedr;
 (function (_WorldSeedr) {
     "use strict";
@@ -12707,14 +13397,9 @@ var WorldSeedr;
 /// <reference path="StringFilr-0.2.1.ts" />
 /// <reference path="ThingHittr-0.2.0.ts" />
 /// <reference path="TimeHandlr-0.2.0.ts" />
+/// <reference path="TouchPassr-0.2.0.ts" />
 /// <reference path="WorldSeedr-0.2.0.ts" />
 /// <reference path="js_beautify.ts" />
-var __extends = this.__extends || function (d, b) {
-    for (var p in b) if (b.hasOwnProperty(p)) d[p] = b[p];
-    function __() { this.constructor = d; }
-    __.prototype = b.prototype;
-    d.prototype = new __();
-};
 var GameStartr;
 (function (_GameStartr) {
     "use strict";
@@ -12772,6 +13457,7 @@ var GameStartr;
                 "resetMapsCreator",
                 "resetMapsHandler",
                 "resetInputWriter",
+                "resetTouchPasser",
                 "resetLevelEditor",
                 "resetWorldSeeder",
                 "resetModAttacher",
@@ -12797,21 +13483,20 @@ var GameStartr;
          * @param {GameStartr} GameStarter
          * @param {Object} [customs]
          * @return {Array} How long each reset Function took followed by the entire
-         * operation, in milliseconds.
+         *                 operation, in milliseconds.
          */
         GameStartr.prototype.resetTimed = function (GameStarter, customs) {
             return EightBittr.EightBittr.prototype.resetTimed(GameStarter, GameStarter.resets, customs);
         };
         /**
-         * Sets self.ObjectMaker.
+         * Sets this.ObjectMaker.
          *
          * Because many Thing functions require access to other FSM modules, each is
          * given a reference to this container FSM via properties.thing.GameStarter.
          *
          * @param {GameStartr} GameStarter
          * @param {Object} [customs]
-         * @remarks Requirement(s): ObjectMakr (src/ObjectMakr/ObjectMakr.js)
-         *                          objects.js (settings/objects.js)
+         * @remarks Requirement(s): objects.js (settings/objects.js)
          */
         GameStartr.prototype.resetObjectMaker = function (GameStarter, customs) {
             GameStarter.ObjectMaker = new ObjectMakr.ObjectMakr(GameStarter.proliferate({
@@ -12828,12 +13513,11 @@ var GameStartr;
             }, GameStarter.settings.objects));
         };
         /**
-         * Sets self.QuadsKeeper.
+         * Sets this.QuadsKeeper.
          *
          * @param {GameStartr} GameStarter
          * @param {Object} [customs]
-         * @remarks Requirement(s): QuadsKeepr (src/QuadsKeepr/QuadsKeepr.js)
-         *                          quadrants.js (settings/quadrants.js)
+         * @remarks Requirement(s): quadrants.js (settings/quadrants.js)
          */
         GameStartr.prototype.resetQuadsKeeper = function (GameStarter, customs) {
             var quadrantWidth = customs.width / (GameStarter.settings.quadrants.numCols - 3), quadrantHeight = customs.height / (GameStarter.settings.quadrants.numRows - 2);
@@ -12849,12 +13533,11 @@ var GameStartr;
             }, GameStarter.settings.quadrants));
         };
         /**
-         * Sets self.PixelRender.
+         * Sets this.PixelRender.
          *
          * @param {GameStartr} GameStarter
          * @param {Object} [customs]
-         * @remarks Requirement(s): PixelRendr (src/PixelRendr/PixelRendr.js)
-         *                          sprites.js (settings/sprites.js)
+         * @remarks Requirement(s): sprites.js (settings/sprites.js)
          */
         GameStartr.prototype.resetPixelRender = function (GameStarter, customs) {
             GameStarter.PixelRender = new PixelRendr.PixelRendr(GameStarter.proliferate({
@@ -12864,12 +13547,11 @@ var GameStartr;
             }, GameStarter.settings.sprites));
         };
         /**
-         * Sets self.PixelDrawer.
+         * Sets this.PixelDrawer.
          *
          * @param {GameStartr} GameStarter
          * @param {Object} [customs]
-         * @remarks Requirement(s): PixelDrawr (src/PixelDrawr/PixelDrawr.js)
-         *                          renderer.js (settings/renderer.js)
+         * @remarks Requirement(s): renderer.js (settings/renderer.js)
          */
         GameStartr.prototype.resetPixelDrawer = function (GameStarter, customs) {
             GameStarter.PixelDrawer = new PixelDrawr.PixelDrawr(GameStarter.proliferate({
@@ -12885,8 +13567,7 @@ var GameStartr;
          *
          * @param {GameStartr} GameStarter
          * @param {Object} [customs]
-         * @remarks Requirement(s): TimeHandlr (src/TimeHandlr/TimeHandlr.js)
-         *                          events.js (settings/events.js)
+         * @remarks Requirement(s): events.js (settings/events.js)
          */
         GameStartr.prototype.resetTimeHandler = function (GameStarter, customs) {
             GameStarter.TimeHandler = new TimeHandlr.TimeHandlr(GameStarter.proliferate({
@@ -12895,12 +13576,11 @@ var GameStartr;
             }, GameStarter.settings.events));
         };
         /**
-         * Sets self.AudioPlayer.
+         * Sets this.AudioPlayer.
          *
          * @param {GameStartr} GameStarter
          * @param {Object} [customs]
-         * @remarks Requirement(s): AudioPlayr (src/AudioPlayr/AudioPlayr.js)
-         *                          audio.js (settings/audio.js)
+         * @remarks Requirement(s): audio.js (settings/audio.js)
          */
         GameStartr.prototype.resetAudioPlayer = function (GameStarter, customs) {
             GameStarter.AudioPlayer = new AudioPlayr.AudioPlayr(GameStarter.proliferate({
@@ -12908,12 +13588,11 @@ var GameStartr;
             }, GameStarter.settings.audio));
         };
         /**
-         * Sets self.GamesRunner.
+         * Sets this.GamesRunner.
          *
          * @param {GameStartr} GameStarter
          * @param {Object} [customs]
-         * @remarks Requirement(s): GamesRunnr (src/GamesRunnr/GamesRunnr.js)
-         *                          runner.js (settings/runner.js)
+         * @remarks Requirement(s): runner.js (settings/runner.js)
          */
         GameStartr.prototype.resetGamesRunner = function (GameStarter, customs) {
             GameStarter.GamesRunner = new GamesRunnr.GamesRunnr(GameStarter.proliferate({
@@ -12926,12 +13605,11 @@ var GameStartr;
             }, GameStarter.settings.runner));
         };
         /**
-         * Sets self.ItemsHolder.
+         * Sets this.ItemsHolder.
          *
          * @param {GameStartr} GameStarter
          * @param {Object} [customs]
-         * @remarks Requirement(s): ItemsHoldr (src/ItemsHoldr/ItemsHoldr.js)
-         *                          statistics.js (settings/statistics.js)
+         * @remarks Requirement(s): statistics.js (settings/statistics.js)
          */
         GameStartr.prototype.resetItemsHolder = function (GameStarter, customs) {
             GameStarter.ItemsHolder = new ItemsHoldr.ItemsHoldr(GameStarter.proliferate({
@@ -12939,23 +13617,21 @@ var GameStartr;
             }, GameStarter.settings.statistics));
         };
         /**
-         * Sets self.GroupHolder.
+         * Sets this.GroupHolder.
          *
          * @param {GameStartr} GameStarter
          * @param {Object} [customs]
-         * @remarks Requirement(s): GroupHoldr (src/ThingHittr/GroupHoldr.js)
-         *                          groups.js (settings/groups.js)
+         * @remarks Requirement(s): groups.js (settings/groups.js)
          */
         GameStartr.prototype.resetGroupHolder = function (GameStarter, customs) {
             GameStarter.GroupHolder = new GroupHoldr.GroupHoldr(GameStarter.settings.groups);
         };
         /**
-         * Sets self.ThingHitter.
+         * Sets this.ThingHitter.
          *
          * @param {GameStartr} GameStarter
          * @param {Object} [customs]
-         * @remarks Requirement(s): ThingHittr (src/ThingHittr/ThingHittr.js)
-         *                          collisions.js (settings/collisions.js)
+         * @remarks Requirement(s): collisions.js (settings/collisions.js)
          */
         GameStartr.prototype.resetThingHitter = function (GameStarter, customs) {
             GameStarter.ThingHitter = new ThingHittr.ThingHittr(GameStarter.proliferate({
@@ -12963,12 +13639,11 @@ var GameStartr;
             }, GameStarter.settings.collisions));
         };
         /**
-         * Sets self.MapScreener.
+         * Sets this.MapScreener.
          *
          * @param {GameStartr} GameStarter
          * @param {Object} [customs]
-         * @remarks Requirement(s): MapScreenr (src/MapScreenr/MapScreenr.js)
-         *                          maps.js (settings/maps.js)
+         * @remarks Requirement(s): maps.js (settings/maps.js)
          */
         GameStartr.prototype.resetMapScreener = function (GameStarter, customs) {
             GameStarter.MapScreener = new MapScreenr.MapScreenr({
@@ -12981,21 +13656,19 @@ var GameStartr;
             });
         };
         /**
-         * Sets self.NumberMaker.
+         * Sets this.NumberMaker.
          *
          * @param {GameStartr} GameStarter
          * @param {Object} [customs]
-         * @remarks Requirement(s): NumberMaker (src/NumberMakr/NumberMakr.js)
          */
         GameStartr.prototype.resetNumberMaker = function (GameStarter, customs) {
             GameStarter.NumberMaker = new NumberMakr.NumberMakr();
         };
         /**
-         * Sets self.MapCreator.
+         * Sets this.MapCreator.
          *
          * @param {GameStartr} GameStarter
-         * @remarks Requirement(s): MapCreatr (src/MapCreatr/MapCreatr.js)
-         *                          maps.js (settings/maps.js)
+         * @remarks Requirement(s): maps.js (settings/maps.js)
          */
         GameStartr.prototype.resetMapsCreator = function (GameStarter, customs) {
             GameStarter.MapsCreator = new MapsCreatr.MapsCreatr({
@@ -13008,12 +13681,11 @@ var GameStartr;
             });
         };
         /**
-         * Sets self.MapsHandler.
+         * Sets this.MapsHandler.
          *
          * @param {GameStartr} GameStarter
          * @param {Object} [customs]
-         * @remarks Requirement(s): MapsHandlr (src/MapsHandlr/MapsHandlr.js)
-         *                          maps.js (settings/maps.js)
+         * @remarks Requirement(s): maps.js (settings/maps.js)
          */
         GameStartr.prototype.resetMapsHandler = function (GameStarter, customs) {
             GameStarter.MapsHandler = new MapsHandlr.MapsHandlr({
@@ -13025,12 +13697,11 @@ var GameStartr;
             });
         };
         /**
-         * Sets self.InputWriter.
+         * Sets this.InputWriter.
          *
          * @param {GameStartr} GameStarter
          * @param {Object} [customs]
-         * @remarks Requirement(s): InputWritr (src/InputWritr/InputWritr.js)
-         *                          input.js (settings/input.js)
+         * @remarks Requirement(s): input.js (settings/input.js)
          */
         GameStartr.prototype.resetInputWriter = function (GameStarter, customs) {
             GameStarter.InputWriter = new InputWritr.InputWritr(GameStarter.proliferate({
@@ -13038,12 +13709,23 @@ var GameStartr;
             }, GameStarter.settings.input.InputWritrArgs));
         };
         /**
-         * Sets self.LevelEditor.
+         * Sets this.InputWriter.
          *
          * @param {GameStartr} GameStarter
          * @param {Object} [customs]
-         * @remarks Requirement(s): LevelEditr (src/LevelEditr/LevelEditr.js)
-         *                          editor.js (settings/editor.js)
+         * @remarks Requirement(s): touch.js (settings/touch.js)
+         */
+        GameStartr.prototype.resetTouchPasser = function (GameStarter, customs) {
+            GameStarter.TouchPasser = new TouchPassr.TouchPassr(GameStarter.proliferate({
+                "InputWriter": GameStarter.InputWriter
+            }, GameStarter.settings.touch));
+        };
+        /**
+         * Sets this.LevelEditor.
+         *
+         * @param {GameStartr} GameStarter
+         * @param {Object} [customs]
+         * @remarks Requirement(s): editor.js (settings/editor.js)
          */
         GameStartr.prototype.resetLevelEditor = function (GameStarter, customs) {
             GameStarter.LevelEditor = new LevelEditr.LevelEditr(GameStarter.proliferate({
@@ -13052,12 +13734,11 @@ var GameStartr;
             }, GameStarter.settings.editor));
         };
         /**
-         * Sets self.WorldSeeder.
+         * Sets this.WorldSeeder.
          *
          * @param {GameStartr} GameStarter
          * @param {Object} [customs]
-         * @remarks Requirement(s): WorldSeedr (src/WorldSeedr/WorldSeedr.js)
-         *                          generator.js (settings/generator.js)
+         * @remarks Requirement(s): generator.js (settings/generator.js)
          */
         GameStartr.prototype.resetWorldSeeder = function (GameStarter, customs) {
             GameStarter.WorldSeeder = new WorldSeedr.WorldSeedr(GameStarter.proliferate({
@@ -13066,12 +13747,11 @@ var GameStartr;
             }, GameStarter.settings.generator));
         };
         /**
-         * Sets self.ModAttacher.
+         * Sets this.ModAttacher.
          *
          * @param {GameStartr} GameStarter
          * @param {Object} [customs]
-         * @remarks Requirement(s): ModAttachr (src/ModAttachr/ModAttachr.js)
-         *                          mods.js (settings/mods.js)
+         * @remarks Requirement(s): mods.js (settings/mods.js)
          */
         GameStartr.prototype.resetModAttacher = function (GameStarter, customs) {
             GameStarter.ModAttacher = new ModAttachr.ModAttachr(GameStarter.proliferate({
