+++ Source/index.html
@@ -67,7 +67,6 @@
     /build -->
     <!-- build:remove -->
     <script src="References/UserWrappr-0.2.0.js" type="text/javascript"></script>
-    <script src="References/TouchPassr-0.0.0.js" type="text/javascript"></script>
     <script src="FullScreenMario.js" type="text/javascript"></script>
     <script src="settings/audio.js" type="text/javascript"></script>
     <script src="settings/collisions.js" type="text/javascript"></script>
